django_backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Database
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0  # PostgreSQL

# Security
python-dotenv>=1.0,<2.0
cryptography>=42.0,<43.0

# Email (production: Resend SDK)
resend>=2.0,<3.0

# API / validation
pydantic>=2.5,<3.0

# HTTP client (used by check-understanding to call AI engine)
httpx>=0.27,<1.0

# HTTP client (used by grading services for GitHub API + ai_engine fan-out)
requests>=2.31,<3.0

# Utils
python-dateutil>=2.8,<3.0

# Development
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0

# Production
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0
